Fix unused/sign warnings
```
tests/libc++/algorithm/any_of.cc:19:34: warning: unused parameter 'x'
tests/libc++/algorithm/all_of.cc:19:33: warning: unused parameter 'x'
tests/libc++/algorithm/count_if.cc:25:35: warning: unused parameter 'x'
tests/libc++/algorithm/none_of.cc:19:34: warning: unused parameter 'x'
tests/tty/string_puts.cc:81:25: warning: comparison of integers of different signs: 'int' and 'const size_t'
tests/tty/string_puts.cc:101:25: warning: comparison of integers of different signs: 'int' and 'const size_t'
tests/scheduler/KillTask.cc:56:13: warning: unused variable 'exitedBefore'
```

+++ tests/tty/string_puts.cc
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <doctest/doctest.h>
 #include <stdint.h>
 #include <string>
@@ -78,7 +79,7 @@ TEST_CASE("putLine_fills_entire_row")
   uint8_t attr = vgaColor(COLOR_WHITE, COLOR_RED);
   CHECK(VGA_RAM[2 * VGA_WIDTH + 0] == vgaEntry('a', attr));
   CHECK(VGA_RAM[2 * VGA_WIDTH + 1] == vgaEntry('b', attr));
-  for (int col = 2; col < VGA_WIDTH; ++col) {
+  for (size_t col = 2; col < VGA_WIDTH; ++col) {
     CHECK(VGA_RAM[2 * VGA_WIDTH + col] == vgaEntry(' ', attr));
   }
 }
@@ -98,7 +99,7 @@ TEST_CASE("putLine_empty_row_clears_row")
   Tty::setColor(vgaColor(COLOR_LIGHT_GREEN, COLOR_BLACK));
   Tty::putLine(string(""), 0);
 
-  for (int col = 0; col < VGA_WIDTH; ++col) {
+  for (size_t col = 0; col < VGA_WIDTH; ++col) {
     CHECK(VGA_RAM[0 * VGA_WIDTH + col] == vgaEntry(' ', vgaColor(COLOR_LIGHT_GREEN, COLOR_BLACK)));
   }
 }
